reorganize link-fetching functions
