+++ backend/app/get_travel_time.py
@@ -1,7 +1,7 @@
 """Function for returning data from the aggregate-travel-times/ endpoint"""
 
 from app.db import getConnection
-from app.get_here_links import get_here_links
+from app.links.here import get_here_links
 from app.selectMapVersion import selectMapVersion
 from traveltimetools.utils import timeFormats
 import numpy
